use class names as file names on oauth packages
